+++ next/src/components/variables/edit/EditVariableForm.tsx
@@ -76,6 +76,7 @@ export default function EditVariableForm({
       submitLabel={t('common.edit')}
       scopes={scopes}
       variables={variables}
+      isExternalDatatypeTypeNameEditable={true}
     />
   );
 }