+++ next/src/components/variables/form/VariableForm.tsx
@@ -30,6 +30,8 @@ type Props = {
   scopes: Map<string, string>;
   /** List of variables used for auto-completion in VTL editor. */
   variables?: Variable[];
+  /** Disable datatype selection for external variables */
+  isExternalDatatypeTypeNameEditable?: boolean;
 };
 
 /**
@@ -54,6 +56,7 @@ export default function VariableForm({
   submitLabel,
   scopes,
   variables = [],
+  isExternalDatatypeTypeNameEditable = false,
 }: Readonly<Props>) {
   const navigate = useNavigate();
 
@@ -71,6 +74,21 @@ export default function VariableForm({
 
   const selectedType = watch('type');
   const selectedTypeName = watch('datatype.typeName');
+
+  const isDatatypeTypeNameDisabled =
+    selectedType === VariableType.External &&
+    !isExternalDatatypeTypeNameEditable;
+
+  const datatypeTypeNameOptions = (() => {
+    // For External variable, enable only current saved option and 'Text'
+    if (selectedType === VariableType.External) {
+      return datatypeOptions.filter(
+        ({ value }) =>
+          value === DatatypeType.Text || value === variable.datatype.typeName,
+      );
+    }
+    return datatypeOptions;
+  })();
 
   /** Ignore dirty state and return to the variables page. */
   const handleCancel = () => {
@@ -271,9 +289,10 @@ export default function VariableForm({
             touched={isTouched}
           >
             <Select<DatatypeType>
-              options={datatypeOptions}
+              options={datatypeTypeNameOptions}
               value={value}
               onChange={onChange}
+              disabled={isDatatypeTypeNameDisabled}
             />
           </Field>
         )}